generate_video_captions/requirements.txt
openai-whisper
ffmpeg-python 